Add comprehensive tests for user flows, performance, load, and property validation
- Implemented end-to-end tests covering user signup, login, search, conversation, messaging, and logout in `e2e_test.rs`.
- Created performance tests measuring message delivery latency, throughput, WebSocket handshake latency, and concurrent connections in `performance_test.rs`.
- Developed a load testing configuration using Locust to simulate 100 concurrent users performing various chat operations in `locustfile.py`.
- Added requirements for load testing in `requirements.txt`.
- Introduced property-based tests for username, password, and message content validation in `property_tests.rs`.

+++ src/backend/handlers/handshake.rs
@@ -102,7 +102,7 @@ impl HandshakeValidator {
 
         // Check token expiration explicitly
         let now = chrono::Utc::now().timestamp();
-        if claims.exp <= now as i64 {
+        if claims.exp <= now {
             return Err((StatusCode::UNAUTHORIZED, "Token has expired".to_string()));
         }
